requirements.txt
openai==0.28.0 # Downgraded from 1.63.0 to fix lambda - pydantic module import error
python-dotenv==1.0.1
flask==3.1.0
aws-wsgi==0.2.7
async_timeout==4.0.2 # for lambda runtime failure fix [needed for openai versions < 1.0]